Fixed issue with copying images and taking photos in api 24

+++ app/src/main/java/me/anon/grow/fragment/PlantDetailsFragment.java
@@ -8,6 +8,7 @@ import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.app.ProgressDialog;
+import android.content.ContentResolver;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -23,6 +24,7 @@ import android.os.ParcelFileDescriptor;
 import android.preference.PreferenceManager;
 import android.provider.MediaStore;
 import android.support.annotation.Nullable;
+import android.support.v4.content.FileProvider;
 import android.text.Html;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
@@ -56,6 +58,7 @@ import java.util.Date;
 import java.util.Locale;
 
 import me.anon.grow.AddWateringActivity;
+import me.anon.grow.BuildConfig;
 import me.anon.grow.EditWateringActivity;
 import me.anon.grow.EventsActivity;
 import me.anon.grow.MainApplication;
@@ -316,28 +319,38 @@ public class PlantDetailsFragment extends Fragment
 						catch (IOException e){}
 
 						File out = new File(path, System.currentTimeMillis() + ".jpg");
+						Uri photoURI = FileProvider.getUriForFile(getActivity(), BuildConfig.APPLICATION_ID + ".provider", out);
 
 						plant.getImages().add(out.getAbsolutePath());
 
-						intent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(out));
+						intent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
 						startActivityForResult(intent, 1);
 					}
 					else
 					{
-						if (Build.VERSION.SDK_INT < 19)
+						if (!PermissionHelper.hasPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE))
 						{
-							Intent intent = new Intent();
-							intent.setType("image/jpeg");
-							intent.setAction(Intent.ACTION_GET_CONTENT);
-							startActivityForResult(Intent.createChooser(intent, "Select picture"), 3);
+							PermissionHelper.doPermissionCheck(PlantDetailsFragment.this, Manifest.permission.WRITE_EXTERNAL_STORAGE, 1, "Need permission");
+							return;
+						}
+
+						Intent intent;
+						if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT)
+						{
+							intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
+							intent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
+							intent.addFlags(Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION);
 						}
 						else
 						{
-							Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
-							intent.addCategory(Intent.CATEGORY_OPENABLE);
-							intent.setType("image/jpeg");
-							startActivityForResult(Intent.createChooser(intent, "Select picture"), 3);
+							intent = new Intent(Intent.ACTION_GET_CONTENT);
 						}
+
+						intent.putExtra(Intent.EXTRA_LOCAL_ONLY, true);
+						intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+						intent.setType("image/*");
+
+						startActivityForResult(Intent.createChooser(intent, "Select picture"), 3);
 					}
 				}
 			})
@@ -411,39 +424,47 @@ public class PlantDetailsFragment extends Fragment
 		}
 		else if (requestCode == 3) // choose image from gallery
 		{
-			Uri selectedUri = data.getData();
-			if (Build.VERSION.SDK_INT >= 19)
+			if (resultCode != Activity.RESULT_CANCELED)
 			{
-				final int takeFlags = data.getFlags() & (Intent.FLAG_GRANT_READ_URI_PERMISSION | Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
-				getActivity().getContentResolver().takePersistableUriPermission(selectedUri, takeFlags);
-			}
+				if (data != null && data.getData() != null)
+				{
+					Uri selectedUri = data.getData();
+					if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT)
+					{
+						getActivity().grantUriPermission(getActivity().getPackageName(), selectedUri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
 
-			File path = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM).getPath() + "/GrowTracker/" + plant.getId() + "/");
-			path.mkdirs();
+						final int takeFlags = data.getFlags() & (Intent.FLAG_GRANT_READ_URI_PERMISSION);
+						getActivity().getContentResolver().takePersistableUriPermission(selectedUri, takeFlags);
+					}
 
-			try
-			{
-				new File(path, ".nomedia").createNewFile();
-			}
-			catch (IOException e){}
+					File path = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM).getPath() + "/GrowTracker/" + plant.getId() + "/");
+					path.mkdirs();
 
-			File out = new File(path, System.currentTimeMillis() + ".jpg");
+					try
+					{
+						new File(path, ".nomedia").createNewFile();
+					}
+					catch (IOException e){}
 
-			copyImage(selectedUri, out);
+					File out = new File(path, System.currentTimeMillis() + ".jpg");
 
-			if (out.exists() && out.length() > 0)
-			{
-				plant.getImages().add(out.getAbsolutePath());
-				PlantManager.getInstance().upsert(plantIndex, plant);
-			}
-			else
-			{
-				out.delete();
+					copyImage(selectedUri, out);
+
+					if (out.exists() && out.length() > 0)
+					{
+						plant.getImages().add(out.getAbsolutePath());
+						PlantManager.getInstance().upsert(plantIndex, plant);
+					}
+					else
+					{
+						out.delete();
+					}
+				}
 			}
 		}
 
 		// both photo options
-		if (requestCode == 1 || requestCode == 3)
+		if ((requestCode == 1 || requestCode == 3) && resultCode != Activity.RESULT_CANCELED)
 		{
 			if (getActivity() != null)
 			{
